Restructure stdsync module
